move addon to separate file
